Added sync break using built-in brk option in serialport
as per LIN messaging standard, and as also mentioned in one of the patents, a sync break must precede each packet to be sent. It consists of sending a series of at least 13 null bits. It can be achieved by either sending a single 0x00 byte at a slower baudrate as in hormann-hcp repo, or just setting the break signal for the appropriate duration
refs:
https://en.wikipedia.org/wiki/Local_Interconnect_Network#Header
https://serialport.io/docs/api-stream#set
https://github.com/raintonr/hormann-hcp/issues/5
https://patents.google.com/patent/WO2005076529A1/en